added a couple more modules
