+++ diet.html
@@ -79,6 +79,7 @@
         </div>
     </footer>
 </body>
-<script src="./main.js/modules.js" type="module"></script>
+<script src="./main.js/displayDataModule.js" type="module"></script>
+<script src="./main.js/errorHandleModule.js" type="module"></script>
 <script src="./main.js/diet.js" type="module"></script>
 </html>
