+++ index.html
@@ -80,6 +80,7 @@
         </div>
     </footer>
 </body>
-<script src="./main.js/modules.js" type="module"></script>
+<script src="./main.js/displayDataModule.js" type="module"></script>
+<script src="./main.js/errorHandleModule.js" type="module"></script>
 <script src="./main.js/script.js" type="module"></script>
 </html>
