+++ main.js/search.js
@@ -1,4 +1,6 @@
-import displaydata from "./modules.js";
+import { displaydata, displayIngredients } from "./displayDataModule.js";
+import errorHandle from "./errorHandleModule.js";
+
 
 
 // Creating Namespace
@@ -42,7 +44,7 @@ recipesApp.apiCallSearch = function(searchedRecipe) {
             recipesApp.apiCallRecipe(jsonData[0].id);
         })
         .catch(function (error) {
-            recipesApp.errorHandle();
+            errorHandle();
         })    
 }
 
@@ -72,50 +74,24 @@ recipesApp.apiCallRecipe = function (recipeIdNumber) {
         }
     })
     .then(function (jsonData) {
-            recipesApp.displayIngredients(jsonData);
+            displayIngredients(jsonData);
             displaydata(jsonData);
             recipesApp.cardClick.style.display = 'block';
         })
         .catch(function (error) {
-            recipesApp.errorHandle();
+            errorHandle();
         })
-}
-
-// Defining a method to get the randomly selected recipe and return the ingredients list
-recipesApp.displayIngredients = function(recipesData) {
-    const ulItem = document.querySelector("#ingredientsList");
-    ulItem.innerHTML = '';
-    const ingredientsArray = recipesData.extendedIngredients;
-    ingredientsArray.forEach( ingredientObj => {
-        const listItem = document.createElement("li");
-        listItem.textContent = ingredientObj.original;
-        ulItem.appendChild(listItem);
-
-    });
-}
-
-// Error Handling no response from API
-recipesApp.errorHandle = function () {
-    const outerDiv = document.querySelector('#recipeResult');
-    outerDiv.innerHTML = '';
-    const ulItem = document.querySelector("#ingredientsList");
-    ulItem.innerHTML = '';
-    const newHeading = document.createElement('h2');
-    newHeading.textContent = 'Sorry - No Recipes match that selection. Please select again';
-    outerDiv.appendChild(newHeading);
-    // add a function here for a random joke when error handling!
-
 }
 
 recipesApp.cardListener = function() {
     recipesApp.cardClick.addEventListener('click', function (event) {
-        const innerCard = document.querySelector('#cardHolderInner').classList.toggle('rotateCard');
+        document.querySelector('#cardHolderInner').classList.toggle('rotateCard');
     })
 }
 // Init
 recipesApp.init = function(){
-    recipesApp.cardListener();
     recipesApp.userSearch(recipesApp.recipeSearch);
+    recipesApp.cardListener();
 }
 // Call the init method to kickstart the app upon page load
 recipesApp.init();
