+++ search.html
@@ -77,6 +77,7 @@
         </div>
     </footer>
 </body>
-<script src="./main.js/modules.js" type="module"></script>
+<script src="./main.js/displayDataModule.js" type="module"></script>
+<script src="./main.js/errorHandleModule.js" type="module"></script>
 <script src="./main.js/search.js" type="module"></script>
 </html>
